requirements.txt
dash
plotly
yfinance
pandas
pandas-ta
openai
google-generativeai
